requirements.txt
Jinja2
Markdown
python-frontmatter
PyYAML
pictex